unint64 replaced with uint for convenience

+++ src/models/torrent.go
@@ -1,10 +1,10 @@
 package models
 
 type Torrent struct {
-	ID uint64
+	ID uint
 	Title string
-	Likes uint64
-	Dislikes uint64
+	Likes uint
+	Dislikes uint
 	URI string
 	Image string
 	Category string
